+++ lib/file-source/src/lib.rs
@@ -5,21 +5,20 @@ extern crate scan_fmt;
 #[macro_use]
 extern crate tracing;
 
+pub mod buffer;
 mod checkpointer;
 mod file_server;
 mod file_watcher;
 mod fingerprinter;
 mod internal_events;
 mod metadata_ext;
-#[cfg(test)]
-mod model_tests;
 pub mod paths_provider;
 
 pub use self::file_server::{FileServer, Shutdown as FileServerShutdown};
 pub use self::fingerprinter::{FingerprintStrategy, Fingerprinter};
 pub use self::internal_events::FileSourceInternalEvents;
 
-type FilePosition = u64;
+pub type FilePosition = u64;
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub enum ReadFrom {
